// src/webscrapping/pc-components/ram/models/crucial.model.ts
export class CrucialRamModel {
  name: string;
  code: string;
  brand: string;
  series: string;
  url: string;
  imageUrl: string;
  price: number;
  type: string;
  moduleType: string;
  capacity: number;
  speed: number;
  voltage: number;
  kitQuantity: number;
  casLatency: number;
  rgb: boolean;
}